fixed cart numerical parser

+++ client/src/components/cart-item/CartItem.js
@@ -8,7 +8,7 @@ const CartItem = ({product, updateQuantity, edit = true, id = false}) => {
             </div>
             <div className={styles['info']}>
                 <div className={styles['title']}>{product.name}</div>
-                <div className={styles['weight']}>{parseInt(product.weight) * parseInt(product.quantity || '1')}{product.measurement}</div>
+                <div className={styles['weight']}>{parseFloat(product.weight) * parseFloat(product.quantity || '1')}{product.measurement}</div>
                 <div className={styles['quantity-wrapper']}>
                     {id && <div className={styles['quantity']}>#{product.product_id}</div>}
                     <div className={styles['quantity']}>{product.quantity? `Quantity: ${product.quantity}` : `Stock:  ${product.stock}`}</div>
@@ -23,7 +23,7 @@ const CartItem = ({product, updateQuantity, edit = true, id = false}) => {
                 </div>
             </div>
             <div className={styles['price']}>
-                <div className={styles['total-price']}>{parseInt(product.price) * parseInt(product.quantity || '1')} EGP</div>
+                <div className={styles['total-price']}>{parseFloat(product.price) * parseFloat(product.quantity || '1')} EGP</div>
                 {parseInt(product.quantity) > 1 && <div
                     className={styles['unit-price']}>{product.price} EGP/{product.weight}{product.measurement}</div>}
             </div>
